# Fixes #31385 - Default to strict qpid-router ciphers