@ekohl 2020-11-25
> While this change only mentions ciphers, in practice it looked like it also disabled TLS 1.0 and 1.1

@pmoravec I'm unsure if it was just nmap's reporting or if the stricter ciphers automatically disabled TLS 1.0 and 1.1. can you shine some light on this?