@pmoravec 2020-12-01
Hi, sorry for later response. I havent tested it by a real client, but running the same `nmap` command and inspecting captured `tcpdump`, I do see that e.g. a Client Hello Handshake with Version TLS1.1 is responded by the router by FIN packet - thus TLS1.1 is disallowed by that ciphers setting. The same happened to TLS1.0 and SSL3.0, all three encryption versions are disallowed, just TLS1.2 is allowed.