@ekohl 2020-12-15
I think that because it was previously undef, it will be updated. However, I don't know that for certain. Perhaps something to try out?