@ehelms 2020-12-17
From my testing, this will need a migration to enable it. If the answer file has an empty answer, it does not treat a new default as the new value. Which I suppose makes sense if you as a user intentionally clear a value hoping for the default or thinking it turns it off perhaps. I am going to merge this as my testing seems to be happy with the change. But please do open a migration for this.