@ekohl 2020-12-18
https://github.com/theforeman/foreman-installer/pull/633 is the migration.